+++ src/configuration/models/stylint/stylintConfigurationRule.ts
@@ -1,0 +1,5 @@
+/**
+ * Model of Stylint Configuration (.stylintrc) file.
+ */
+
+export type StylintConfigurationRule<T> = boolean | T | { expect?: T; error?: boolean };