+++ test/unit/src/commands/clientPackageCommand.spec.ts
@@ -108,6 +108,7 @@ describe("ClientPackageCommand", () => {
             e2eTestFramework: "MochaChai",
             cssPre: "Sass",
             cssPost: "None",
+            cssLinter: "None",
             linter: "ESLint",
             packageManager: "Yarn",
             taskManager: "Gulp",