+++ test/unit/src/commands/platformCommand.spec.ts
@@ -86,6 +86,7 @@ describe("PlatformCommand", () => {
             e2eTestFramework: "MochaChai",
             cssPre: "Sass",
             cssPost: "None",
+            cssLinter: "None",
             linter: "ESLint",
             packageManager: "Yarn",
             taskManager: "Gulp",