+++ test/unit/src/configuration/models/sassLint/sassLintConfiguration.spec.ts
@@ -1,0 +1,12 @@
+/**
+ * Tests for SassLintConfiguration.
+ */
+import * as Chai from "chai";
+import { SassLintConfiguration } from "../../../../../../src/configuration/models/sassLint/sassLintConfiguration";
+
+describe("SassLintConfiguration", () => {
+    it("can be created", async() => {
+        const obj = new SassLintConfiguration();
+        Chai.should().exist(obj);
+    });
+});