+++ test/unit/src/configuration/models/styleLint/styleLintConfiguration.spec.ts
@@ -1,0 +1,12 @@
+/**
+ * Tests for StyleLintConfiguration.
+ */
+import * as Chai from "chai";
+import { StyleLintConfiguration } from "../../../../../../src/configuration/models/styleLint/styleLintConfiguration";
+
+describe("StyleLintConfiguration", () => {
+    it("can be created", async() => {
+        const obj = new StyleLintConfiguration();
+        Chai.should().exist(obj);
+    });
+});