+++ test/unit/src/configuration/models/stylint/stylintConfiguration.spec.ts
@@ -1,0 +1,12 @@
+/**
+ * Tests for StylintConfiguration.
+ */
+import * as Chai from "chai";
+import { StylintConfiguration } from "../../../../../../src/configuration/models/stylint/stylintConfiguration";
+
+describe("StylintConfiguration", () => {
+    it("can be created", async() => {
+        const obj = new StylintConfiguration();
+        Chai.should().exist(obj);
+    });
+});