+++ test/unit/src/pipelineSteps/cssLinter/sassLint.spec.ts
@@ -1,0 +1,147 @@
+/**
+ * Tests for SassLint.
+ */
+import * as Chai from "chai";
+import * as Sinon from "sinon";
+import { IFileSystem } from "unitejs-framework/dist/interfaces/IFileSystem";
+import { ILogger } from "unitejs-framework/dist/interfaces/ILogger";
+import { SassLintConfiguration } from "../../../../../src/configuration/models/sassLint/sassLintConfiguration";
+import { UniteConfiguration } from "../../../../../src/configuration/models/unite/uniteConfiguration";
+import { EngineVariables } from "../../../../../src/engine/engineVariables";
+import { SassLint } from "../../../../../src/pipelineSteps/cssLinter/sassLint";
+import { FileSystemMock } from "../../fileSystem.mock";
+
+describe("SassLint", () => {
+    let sandbox: Sinon.SinonSandbox;
+    let loggerStub: ILogger;
+    let loggerErrorSpy: Sinon.SinonSpy;
+    let fileSystemMock: IFileSystem;
+    let uniteConfigurationStub: UniteConfiguration;
+    let engineVariablesStub: EngineVariables;
+
+    beforeEach(() => {
+        sandbox = Sinon.sandbox.create();
+        loggerStub = <ILogger>{};
+        loggerStub.info = () => { };
+        loggerStub.error = () => { };
+        loggerErrorSpy = sandbox.spy(loggerStub, "error");
+
+        fileSystemMock = new FileSystemMock();
+        uniteConfigurationStub = new UniteConfiguration();
+        uniteConfigurationStub.cssLinter = "SassLint";
+        uniteConfigurationStub.cssPre = "Sass";
+
+        engineVariablesStub = new EngineVariables();
+        engineVariablesStub.setupDirectories(fileSystemMock, "./test/unit/temp");
+        engineVariablesStub.findDependencyVersion = sandbox.stub().returns("1.2.3");
+    });
+
+    afterEach(async () => {
+        sandbox.restore();
+        await fileSystemMock.directoryDelete("./test/unit/temp");
+    });
+
+    it("can be created", () => {
+        const obj = new SassLint();
+        Chai.should().exist(obj);
+    });
+
+    describe("mainCondition", () => {
+        it("can be called with not matching condition", async () => {
+            const obj = new SassLint();
+            uniteConfigurationStub.cssLinter = undefined;
+            const res = obj.mainCondition(uniteConfigurationStub, engineVariablesStub);
+            Chai.expect(res).to.be.equal(false);
+        });
+
+        it("can be called with matching condition", async () => {
+            const obj = new SassLint();
+            const res = obj.mainCondition(uniteConfigurationStub, engineVariablesStub);
+            Chai.expect(res).to.be.equal(true);
+        });
+    });
+
+    describe("intitialise", () => {
+        it("can be called with false main condition", async () => {
+            const obj = new SassLint();
+            const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, false);
+            Chai.expect(res).to.be.equal(0);
+            Chai.expect(engineVariablesStub.getConfiguration("SassLint")).to.be.equal(undefined);
+        });
+
+        it("can be called with mismatched cssPre", async () => {
+            uniteConfigurationStub.cssPre = "Less";
+            const obj = new SassLint();
+            const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
+            Chai.expect(res).to.be.equal(1);
+            Chai.expect(engineVariablesStub.getConfiguration("SassLint")).to.be.equal(undefined);
+            Chai.expect(loggerErrorSpy.args[0][0]).contains("Sass");
+        });
+
+        it("can succeed when file does exist", async () => {
+            fileSystemMock.fileExists = sandbox.stub().onFirstCall().resolves(true);
+            fileSystemMock.fileReadJson = sandbox.stub().resolves({ options: { "merge-default-rules": false } });
+            const obj = new SassLint();
+            const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
+            Chai.expect(res).to.be.equal(0);
+            Chai.expect(engineVariablesStub.getConfiguration<SassLintConfiguration>("SassLint").options["merge-default-rules"]).to.be.equal(false);
+        });
+
+        it("can succeed when file does not exist", async () => {
+            const obj = new SassLint();
+            const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
+            Chai.expect(res).to.be.equal(0);
+            Chai.expect(engineVariablesStub.getConfiguration<SassLintConfiguration>("SassLint").options["merge-default-rules"]).to.be.equal(true);
+        });
+    });
+
+    describe("configure", () => {
+        it("can be called", async () => {
+            const obj = new SassLint();
+            const res = await obj.configure(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
+            Chai.expect(res).to.be.equal(0);
+
+            const packageJsonDevDependencies: { [id: string]: string } = {};
+            engineVariablesStub.buildDevDependencies(packageJsonDevDependencies);
+
+            Chai.expect(packageJsonDevDependencies["sass-lint"]).to.be.equal("1.2.3");
+        });
+
+        it("can be called with false mainCondition", async () => {
+            const obj = new SassLint();
+            const res = await obj.configure(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, false);
+            Chai.expect(res).to.be.equal(0);
+
+            const packageJsonDevDependencies: { [id: string]: string } = { "sass-lint": "1.2.3"};
+            engineVariablesStub.buildDevDependencies(packageJsonDevDependencies);
+
+            Chai.expect(packageJsonDevDependencies["sass-lint"]).to.be.equal(undefined);
+        });
+    });
+
+    describe("finalise", () => {
+        it("can succeed writing", async () => {
+            await fileSystemMock.directoryCreate("./test/unit/temp/www/");
+
+            const obj = new SassLint();
+            await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
+            const res = await obj.finalise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
+            Chai.expect(res).to.be.equal(0);
+
+            const exists = await fileSystemMock.fileExists("./test/unit/temp/www/", ".sasslintrc");
+            Chai.expect(exists).to.be.equal(true);
+        });
+
+        it("can be called with false mainCondition", async () => {
+            await fileSystemMock.directoryCreate("./test/unit/temp/www/");
+            await fileSystemMock.fileWriteJson("./test/unit/temp/www/", ".sasslintrc", {});
+
+            const obj = new SassLint();
+            const res = await obj.finalise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, false);
+            Chai.expect(res).to.be.equal(0);
+
+            const exists = await fileSystemMock.fileExists("./test/unit/temp/www/", ".sasslintrc");
+            Chai.expect(exists).to.be.equal(false);
+        });
+    });
+});